requirements.txt
asgiref==3.2.10
Django==3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
Pillow==7.2.0
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
